Rename `zha.api` to `zha.websocket_api`
